requirements.txt
exifread
requests
os
flask
werkzeug.utils
clarifai.rest